move mrec service to ad wrapper
